Average tier by regions
It gives you unofficial average tier of League of Legends in different regions by using LOL API.
It takes long time to get the number of players in a division in a tier because there are a number of pages consisting of about 205 players in a division and I do not know the exact number! ( Please let me know if there is any way to get the number ) 
The numbers in the csv file indicates the mmr points of each division, from DIAMOND I to IRON IV in order.
if the average tier of a region is 1147, then the region's average tier would be SILVER II 47 points.